requirements.txt
falcon
waitress
Authlib 
requests
falcon-auth
python-dotenv